+++ Sources/SKTilemap.swift
@@ -129,7 +129,19 @@ internal let TileSize8x8   = CGSize(width: 8, height: 8)
 internal let TileSize16x16 = CGSize(width: 16, height: 16)
 internal let TileSize32x32 = CGSize(width: 32, height: 32)
 
-    
+
+/**
+ Delegate for users to implement their own callbacks.
+ */
+public protocol SKTilemapDelegate: class {
+    func didBeginParsing(_ tilemap: SKTilemap)
+    func didAddTileset(_ tileset: SKTileset)
+    func didAddLayer(_ layer: TiledLayerObject)
+    func didReadMap(_ tilemap: SKTilemap)
+    func didRenderMap(_ tilemap: SKTilemap)
+}
+
+
 /**
  The `SKTilemap` class represents a container which manages layers, tiles (sprites), vector objects & images.
  
@@ -164,7 +176,7 @@ open class SKTilemap: SKNode, SKTiledObject {
     open var tilesets: Set<SKTileset> = []                        // tilesets
     
     // current layers
-    private var layers: Set<TiledLayerObject> = []                // layers
+    private var layers: Set<TiledLayerObject> = []                // tile map layers
     open var layerCount: Int { return self.layers.count }         // layer count attribute
     open var properties: [String: String] = [:]                   // custom properties
     open var zDeltaForLayers: CGFloat = 50                        // z-position range for layers
@@ -194,7 +206,10 @@ open class SKTilemap: SKNode, SKTiledObject {
     // dynamics
     open var gravity: CGVector = CGVector.zero
     
-    /// Rendered size of the map in pixels.
+    /// Weak reference to tile map delegate.
+    weak open var delegate: SKTilemapDelegate?
+    
+    /// Rendered size of the map in points.
     open var sizeInPoints: CGSize {
         switch orientation {
         case .orthogonal:
@@ -226,12 +241,12 @@ open class SKTilemap: SKNode, SKTiledObject {
         }
     }
     
-    // returns the last GID for all of the tilesets.
+    /// Returns the last GID for all of the tilesets.
     open var lastGID: Int {
         return tilesets.count > 0 ? tilesets.map {$0.lastGID}.max()! : 0
     }    
     
-    /// Returns the last GID for all tilesets.
+    /// Returns the last index for all tilesets.
     open var lastIndex: Int {
         return layers.count > 0 ? layers.map {$0.index}.max()! : 0
     }
@@ -292,7 +307,7 @@ open class SKTilemap: SKNode, SKTiledObject {
         }
     }
     
-    /// pauses the node, and colors all of its children darker.
+    /// Pauses the node, and colors all of its children darker.
     override open var isPaused: Bool {
         didSet {
             guard oldValue != isPaused else { return }
@@ -311,17 +326,17 @@ open class SKTilemap: SKNode, SKTiledObject {
         }
     }
     
-
     // MARK: - Loading
     
     /**
      Load a Tiled tmx file and return a new `SKTilemap` object. Returns nil if there is a problem reading the file
      
-     - parameter filename: `String` Tiled file name.
+     - parameter filename:   `String` Tiled file name.
+     - parameter delegate:   `SKTilemapDelegate?` optional tilemap delegate instance.
      - returns: `SKTilemap?` tilemap object (if file read succeeds).
      */
-    open class func load(fromFile filename: String) -> SKTilemap? {
-        if let tilemap = SKTilemapParser().load(fromFile: filename) {
+    open class func load(fromFile filename: String, delegate: SKTilemapDelegate? = nil) -> SKTilemap? {
+        if let tilemap = SKTilemapParser().load(fromFile: filename, delegate: delegate) {
             return tilemap
         }
         return nil
@@ -516,7 +531,7 @@ open class SKTilemap: SKNode, SKTiledObject {
         layer.index = layers.count > 0 ? lastIndex + 1 : 0
         
         // setup the layer
-        tileLayerDidBeginRendering(layer: layer)
+        layer.opacity = 0
         
         // don't add the base layer
         if base == false {
@@ -604,7 +619,7 @@ open class SKTilemap: SKNode, SKTiledObject {
      
      - parameter named: `String` layer name.
      */
-    open func isolateLayer(_ named: String?=nil) {
+    open func isolateLayer(_ named: String? = nil) {
         guard named != nil else {
             layers.forEach {$0.visible = true}
             return
@@ -960,23 +975,13 @@ open class SKTilemap: SKNode, SKTiledObject {
     }
     
     // MARK: - Callbacks
-    /**
-     Called when parser begins reading the map.
-     
-     - parameter verbose: `Bool` verbose output.
-     */
-    open func didBeginParsing(verbose: Bool=false, completion: (() -> ())? = nil) {
-        if completion != nil { completion!() }
-    }
     
     /**
      Called when parser has finished reading the map.
      
      - parameter timeStarted: `Date` render start time.
-     - parameter verbose:     `Bool` verbose output.
-     - parameter completion:  `()->()?` optional completion handler.
      */
-    open func didFinishParsing(timeStarted: Date, verbose: Bool=true, completion: (() -> ())? = nil) {
+    open func didFinishRendering(timeStarted: Date) {
         // set the z-depth of the baseLayer
         baseLayer.zPosition = lastZPosition + (zDeltaForLayers * 0.5)
         
@@ -985,38 +990,13 @@ open class SKTilemap: SKNode, SKTiledObject {
         let timeStamp = String(format: "%.\(String(3))f", timeInterval)        
         print("\n -> Success! tile map \"\(name != nil ? name! : "null")\" rendered in: \(timeStamp)s\n")
         
-        // dump the output of the current map to stdout
-        if (verbose == true) {
-            debugLayers()
-        }
-        
         // transfer attributes
         if let scene = scene as? SKTiledScene {
             scene.physicsWorld.gravity = gravity
         }
         
-        // run completion handler
-        if completion != nil { completion!() }
-    }
-    
-    /**
-     Called when parser begins rendering a layer.
-     
-     - parameter layer: `TiledLayerObject` layer instance.
-     */
-    open func tileLayerDidBeginRendering(layer: TiledLayerObject) {
-        layer.opacity = 0
-        layer.isRendered = false
-    }
-    
-    /**
-     Called when parser finishes rendering a layer.
-     
-     - parameter layer:    `TiledLayerObject` layer instance.
-     - parameter duration: `TimeInterval` fade-in duration.
-     */
-    open func tileLayerDidFinishRendering(layer: TiledLayerObject, duration: TimeInterval=0) {
-        layer.parseProperties(completion: nil)
+        // delegate callbacks
+        if self.delegate != nil { delegate!.didRenderMap(self) }
     }
 }
 
@@ -1198,15 +1178,11 @@ extension SKTilemap {
     }
     
     override open var description: String {
-        var tilemapName = "(None)"
-        if let name = name {
-            tilemapName = "\"\(name)\""
-        }
         let renderSizeDesc = "\(sizeInPoints.width.roundTo(1)) x \(sizeInPoints.height.roundTo(1))"
         let sizeDesc = "\(Int(size.width)) x \(Int(size.height))"
         let tileSizeDesc = "\(Int(tileSize.width)) x \(Int(tileSize.height))"
         
-        return "Map: \(tilemapName), \(renderSizeDesc): (\(sizeDesc) @ \(tileSizeDesc))"
+        return "Map: \(name ?? "(None)"), \(renderSizeDesc): (\(sizeDesc) @ \(tileSizeDesc)): \(tileCount) tiles"
     }
     
     override open var debugDescription: String { return description }
@@ -1229,6 +1205,7 @@ extension SKTilemap {
             print("# Tilemap \"\(name != nil ? name! : "null")\": 0 Layers")
             return
         }
+        
         let largestName = layerNames().max() { (a, b) -> Bool in a.characters.count < b.characters.count }
         
         // format the header
@@ -1257,4 +1234,41 @@ extension SKTilemap {
         
         print("\n")
     }
+}
+
+
+/**
+ Add default implementations of callback methods
+ */
+extension SKTilemapDelegate {
+    /**
+     Called when the tilemap is instantiated.
+
+     - parameter tilemap:  `SKTilemap` tilemap instance.
+     */
+    public func didBeginParsing(_ tilemap: SKTilemap) {}
+    /**
+     Called when a tileset is instantiated.
+
+     - parameter tileset:  `SKTileset` tileset instance.
+     */
+    public func didAddTileset(_ tileset: SKTileset) {}
+    /**
+     Called when a layer is added to a tilemap.
+
+     - parameter layer:  `TiledLayerObject` tilemap instance.
+     */
+    public func didAddLayer(_ layer: TiledLayerObject) {}
+    /**
+     Called when the tilemap is finished parsing.
+
+     - parameter tilemap:  `SKTilemap` tilemap instance.
+     */
+    public func didReadMap(_ tilemap: SKTilemap) {}
+    /**
+     Called when the tilemap layers are finished rendering.
+
+     - parameter tilemap:  `SKTilemap` tilemap instance.
+     */
+    public func didRenderMap(_ tilemap: SKTilemap) {}
 }